+++ .gitignore
@@ -1,2 +1,2 @@
-back/node_modules/
-front/node_modules/
+dat-blog-back/node_modules/
+dat-blog-front/node_modules/
