+++ README.md
@@ -1,10 +1,25 @@
 # dat-blog
 Simple blog project to study
 
-Front: 
-cd front
-npm run start
+## Front: 
 
-Back: 
-cd back
+cd dat-blog-front
+yarn start
+
+## Back: 
+
+cd dat-blog-back
 nodemon npm run start
+
+
+
+### Shortcut
+
+install dependencies at front and back end
+./run-prepare.sh 
+
+start front project
+./run-front.sh
+
+start back project
+./run-back.sh
