--- front/src/about/about.jsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default props => (
-    <h1>bdasbd</h1>
-)
