--- front/src/components/template/header.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-import logo from './logo.svg';  
-
-export default props => (
-    <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <h2>asd</h2>
-    </header>
-)
