--- front/src/components/template/menu.jsx
@@ -1,8 +1,0 @@
-import React from 'react'
-
-export default props => (
-    <div>
-        <a href='#/about'>Sobre</a>
-        <a href='#/posts'>Posts</a>
-    </div>
-)
