+++ run-back.sh
@@ -1,0 +1,2 @@
+cd 'dat-blog-back'
+PORT=4002 nodemon npm run start
