+++ run-front.sh
@@ -1,0 +1,2 @@
+cd 'dat-blog-front'
+PORT=4001 yarn start
