+++ run-prepare.sh
@@ -1,0 +1,9 @@
+echo 'installing front end dependencies'
+cd 'dat-blog-front'
+yarn install
+
+cd ..
+
+echo 'installing back end dependencies'
+cd 'dat-blog-back'
+npm install
